Moved models to ensembl_production app
